+++ src/Pages/KelolaKatalog/KelolaKatalog.js
@@ -1,16 +1,21 @@
 import React from 'react'
 import { useEffect, useState } from 'react';
 import { useProSidebar } from 'react-pro-sidebar';
-import { gql, useQuery, useMutation, useLazyQuery } from "@apollo/client"
+import { gql, useQuery, useMutation, useLazyQuery, useSubscription } from "@apollo/client"
 import Modal from 'react-modal';
 
 import Navbar from '../../Components/Navbar/Navbar'
 import Navigation from '../../Components/Sidebar/Sidebar';
+import LoadingSvg from '../../Components/Loading/LoadingSvg';
+
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 import { FaBars } from "react-icons/fa"
 import { AiFillCloseCircle } from "react-icons/ai"
 
 import { GetKatalog } from '../../Graphql/query';
+import { SubscriptionKatalog } from '../../Graphql/subscription';
 import { InsertKatalog } from '../../Graphql/mutation';
 import { DeleteKatalog } from '../../Graphql/mutation';
 import { UpdateKatalog } from '../../Graphql/mutation';
@@ -45,14 +50,18 @@ function KelolaKatalog() {
     width > 1024 && collapseSidebar(false);
   },[width]);
 
-  const {data, loading, error} = useQuery(GetKatalog)
+  const {data, loading, error} = useSubscription(SubscriptionKatalog)
 
   const [katalog, setKatalog] = useState({
     foto: "",
     nama: "", 
     deskripsi: "",
     harga: "",
-    gender: ""
+    gender: "",
+    ukuran: "",
+    kode_produk: "",
+    material: "",
+    stok: 0
   })
 
   const [insertKatalog, {loading: loadingInsertKatalog}] = useMutation(InsertKatalog)
@@ -86,20 +95,40 @@ function KelolaKatalog() {
           harga: katalog.harga,
           deskripsi: katalog.deskripsi,
           gender: katalog.gender,
+          ukuran: katalog.ukuran,
+          kode_produk: katalog.kode_produk,
+          material: katalog.material,
+          stok: katalog.stok,
         }
       }
     })
+    toast.success("Data berhasil ditambahkan")
+    setIsOpenInsert(false)
     // window.location.reload(false);
   }
 
+  const [modalIsOpenDelete, setIsOpenDelete] = useState(false);
+  function openModalDelete() {
+    setIsOpenDelete(true);
+  }
+  function closeModalDelete() {
+    setIsOpenDelete(false);
+  }
+
   const handleDeleteKatalog = (value) => {
-    deleteKatalog({
-      variables: {
-        _eq: value
-      }
-    })
-    alert("Kain Berhasil Dihapus")
-    window.location.reload(false);
+
+    if(window.confirm("Apakah Anda yakin ingin menghapus katalog ini?") == true ) {
+      deleteKatalog({
+        variables: {
+          _eq: value
+        }
+      })
+      toast.success("Data berhasil dihapus")
+      // window.location.reload(false);
+    } else {
+      // window.location.reload(false);
+    }
+    
   }
 
 
@@ -139,15 +168,28 @@ function KelolaKatalog() {
             deskripsi: updateStateKatalog.deskripsi,
             harga: updateStateKatalog.harga,
             nama: updateStateKatalog.nama,
-            gender: updateStateKatalog.gender
+            gender: updateStateKatalog.gender,
+            ukuran: updateStateKatalog.ukuran,
+            kode_produk: updateStateKatalog.kode_produk,
+            material: updateStateKatalog.material,
+            stok: updateStateKatalog.stok,
         }
       }
     })
+    toast.success("Data berhasil diubah")
+    setIsOpenUpdate(false)
   }
   console.log(updateStateKatalog)
 
+  useEffect(() => {
+    console.log("loading", loading)
+
+  }, [loading])
+  
+
   return (
     <div className='flex h-full'>
+      <ToastContainer/>
       <Navigation />
       <main className='w-full'>
         <div className='bg-secondary3 p-5 block md:hidden'>
@@ -158,39 +200,43 @@ function KelolaKatalog() {
             <h2 className='text-4xl lg:text-6xl font-bold text-primary uppercase'>Kelola Katalog</h2>
             <div className='mt-10'>
               <div className='flex justify-end'>
-                <button onClick={openModalInsert} className='bg-secondary px-10 py-3 rounded-md text-white font-medium border border-secondary hover:bg-white hover:text-secondary duration-200'>Tambah Kain</button>
+                <button onClick={openModalInsert} className='bg-secondary px-10 py-3 rounded-md text-white font-medium border border-secondary hover:bg-white hover:text-secondary duration-200'>Tambah Katalog</button>
               </div>
 
-              <table className='table-fixed w-full mt-10 '>
-                <thead>
-                  <tr className='table-fixed'>
-                    <th className='font-semibold uppercase px-5 py-1 text-secondary'>Foto </th>
-                    <th className='font-semibold uppercase px-5 py-1 text-secondary'>Kode Produk</th>
-                    <th className='font-semibold uppercase px-5 py-1 text-secondary'>Nama</th>
-                    <th className='font-semibold uppercase px-5 py-1 text-secondary'>Deskripsi</th>
-                    <th className='font-semibold uppercase px-5 py-1 text-secondary'>Harga</th>
-                    <th className='font-semibold uppercase px-5 py-1 text-secondary'>Aksi</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {data?.sekargaluhetnic_katalog?.map((katalog) => 
-                  <tr className='py-2 border-b'>
-                    <th className='py-2'><img className='w-32 h-32 object-cover' src={katalog.foto}></img></th>
-                    <td className='px-5 py-1'>Kode Produk</td>
-                    <td className='px-5 py-1'>{katalog.nama}</td>
-                    <td className='px-5 py-1'>{katalog.deskripsi}</td>
-                    <th className='px-5 py-1 font-normal'>Rp{katalog.harga.toLocaleString()}</th>
-                    <th className='px-5 py-1'>
-                      <button className='bg-secondary2 text-white px-2 py-1 rounded-md text-sm border border-secondary2 font-light mr-2' onClick={() => {
-                        openModalUpdate();
-                        handleSelectKatalog(katalog.id);}}>Ubah
-                      </button>
-                      <button className='bg-red-700 text-white px-2 py-1 rounded-md border border-red-700 text-sm font-light' onClick={() => handleDeleteKatalog(katalog.id)}>Hapus</button>
-                    </th>
-                  </tr>
-                  )}
-                </tbody>
-              </table>
+
+              { loading ? <LoadingSvg/> :
+                <table className='table-fixed w-full mt-10 '>
+                  <thead>
+                    <tr className='table-fixed'>
+                      <th className='font-semibold uppercase px-5 py-1 text-secondary'>Foto </th>
+                      <th className='font-semibold uppercase px-5 py-1 text-secondary'>Kode Produk</th>
+                      <th className='font-semibold uppercase px-5 py-1 text-secondary'>Nama</th>
+                      <th className='font-semibold uppercase px-5 py-1 text-secondary'>Deskripsi</th>
+                      <th className='font-semibold uppercase px-5 py-1 text-secondary'>Harga</th>
+                      <th className='font-semibold uppercase px-5 py-1 text-secondary'>Aksi</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    {data?.sekargaluhetnic_katalog?.map((katalog) => 
+                        <tr className='py-2 border-b'>
+                          {/* <div className='flex justify-center'>{loading ? <LoadingSvg/> : ""}</div> */}
+                          <th className='py-2'><img className='w-32 h-32 object-cover' src={katalog.foto}></img></th>
+                          <td className='px-5 py-1'>Kode Produk</td>
+                          <td className='px-5 py-1'>{katalog.nama}</td>
+                          <td className='px-5 py-1'>{katalog.deskripsi}</td>
+                          <th className='px-5 py-1 font-normal'>Rp{katalog.harga.toLocaleString()}</th>
+                          <th className='px-5 py-1'>
+                            <button className='bg-secondary2 text-white px-2 py-1 rounded-md text-sm border border-secondary2 font-light mr-2' onClick={() => {
+                              openModalUpdate();
+                              handleSelectKatalog(katalog.id);}}>Ubah
+                            </button>
+                            <button className='bg-red-700 text-white px-2 py-1 rounded-md border border-red-700 text-sm font-light' onClick={() => handleDeleteKatalog(katalog.id)}>Hapus</button>
+                          </th>
+                        </tr>
+                    )}
+                  </tbody>
+                </table>
+              }
             </div>
           </div>
 
@@ -198,9 +244,10 @@ function KelolaKatalog() {
           <Modal
           isOpen={modalIsOpenInsert}
           onRequestClose={closeModalInsert}
-          style={customStyles}
-        >
-          <div className='w-96'>
+          // style={customStyles}
+          className="bg-white shadow-md p-10 w-[550px] mx-auto h-[550px] overflow-y-scroll mt-20 border rounded-md"
+          >
+          <div className='w-96 mx-auto'>
             <AiFillCloseCircle className='w-7 h-7 fill-secondary hover:fill-red-700 duration-200 cursor-pointer float-right' onClick={closeModalInsert}></AiFillCloseCircle>
             <div>
               <h6 className='font-semibold text-lg'>Tambah Kain</h6>
@@ -211,7 +258,7 @@ function KelolaKatalog() {
                 </div>
                 <div className='mb-5'>
                   <p className='text-primary'>Nama</p>
-                  <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='nama' placeholder='Nama Kain'></input>
+                  <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='nama' placeholder='Nama'></input>
                 </div>
                 <div className='mb-5'>
                   <p className='text-primary'>Deksripsi</p>
@@ -219,11 +266,27 @@ function KelolaKatalog() {
                 </div>
                 <div className='mb-5'>
                   <p className='text-primary'>Harga</p>
-                  <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='harga' placeholder='Harga'></input>
+                  <input type='number' onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='harga' placeholder='Harga'></input>
                 </div>
                 <div className='mb-5'>
                   <p className='text-primary'>Gender</p>
                   <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='gender' placeholder='Gender'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Ukuran</p>
+                  <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='ukuran' placeholder='Ukuran'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Kode Produk</p>
+                  <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='kode_produk' placeholder='Kode Produk'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Material</p>
+                  <input onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='material' placeholder='Material'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Stok</p>
+                  <input type='number' onChange={handleInputInsertKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='stok' placeholder='Stok'></input>
                 </div>
                 <div className='flex justify-center mt-10'>
                   <button className='bg-secondary px-10 py-2 rounded-md text-white border border-secondary hover:bg-white hover:text-secondary duration-200 mx-auto' onClick={handleInsertKatalog}>Submit</button>
@@ -238,19 +301,20 @@ function KelolaKatalog() {
         <Modal
           isOpen={modalIsOpenUpdate}
           onRequestClose={closeModalUpdate}
-          style={customStyles}
+          // style={customStyles}
+          className="bg-white shadow-md p-10 w-[550px] mx-auto h-[550px] overflow-y-scroll mt-20 border rounded-md"
         >
-          <div className='w-96'>
+          <div className='w-96 mx-auto'>
             <AiFillCloseCircle className='w-7 h-7 fill-secondary hover:fill-red-700 duration-200 cursor-pointer float-right' onClick={closeModalUpdate}></AiFillCloseCircle>
             <div>
               <h6 className='font-semibold text-lg'>Update Kain</h6>
               <div className='mt-5'>
                 <div className='mb-5'>
-                  <p className='text-primary'>Gambar Kain</p>
+                  <p className='text-primary'>Gambar</p>
                   <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='foto' value={updateStateKatalog.foto} placeholder='Gambar Kain'></input>
                 </div>
                 <div className='mb-5'>
-                  <p className='text-primary'>Nama Kain</p>
+                  <p className='text-primary'>Nama</p>
                   <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='nama' value={updateStateKatalog.nama} placeholder='Nama Kain'></input>
                 </div>
                 <div className='mb-5'>
@@ -259,16 +323,49 @@ function KelolaKatalog() {
                 </div>
                 <div className='mb-5'>
                   <p className='text-primary'>Harga</p>
-                  <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='harga' value={updateStateKatalog.harga} placeholder='Harga'></input>
+                  <input type='number' onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='harga' value={updateStateKatalog.harga} placeholder='Harga'></input>
                 </div>
                 <div className='mb-5'>
                   <p className='text-primary'>Gender</p>
                   <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='gender' value={updateStateKatalog.gender} placeholder='Gender'></input>
                 </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Ukuran</p>
+                  <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='ukuran' value={updateStateKatalog.ukuran} placeholder='Ukuran'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Kode Produk</p>
+                  <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='kode_produk' value={updateStateKatalog.kode_produk} placeholder='Kode Produk'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Material</p>
+                  <input onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='material' value={updateStateKatalog.material} placeholder='Material'></input>
+                </div>
+                <div className='mb-5'>
+                  <p className='text-primary'>Stok</p>
+                  <input type='number' onChange={handleChangeUpdateKatalog} className='border-b focus:outline-none focus:border-primary p-1 text-sm mt-1 w-full' name='stok' value={updateStateKatalog.stok} placeholder='Stok'></input>
+                </div>
                 <div className='flex justify-center mt-10'>
                   <button className='bg-secondary px-10 py-2 rounded-md text-white border border-secondary hover:bg-white hover:text-secondary duration-200 mx-auto' onClick={handleUpdateKatalog}>Submit</button>
                 </div>
               </div>
+            </div>
+
+          </div>
+        </Modal>
+
+
+        <Modal
+          isOpen={modalIsOpenDelete}
+          onRequestClose={closeModalDelete}
+          // style={customStyles}
+          className="bg-white shadow-md p-10 w-[550px] mx-auto h-[550px] overflow-y-scroll mt-20 border rounded-md"
+        >
+          <div className='w-96 mx-auto'>
+            <AiFillCloseCircle className='w-7 h-7 fill-secondary hover:fill-red-700 duration-200 cursor-pointer float-right' onClick={closeModalDelete}></AiFillCloseCircle>
+            <div>
+              <button onClick={() => handleDeleteKatalog(katalog.id)}></button>
+
             </div>
 
           </div>
